Organizes files into domain specific packages
