+++ src/main/java/com/gildedrose/GildedRose.java
@@ -1,5 +1,8 @@
 package com.gildedrose;
 
+import com.gildedrose.goblins_grotto.Item;
+import com.gildedrose.items.*;
+
 import java.util.HashMap;
 
 class GildedRose {
