+++ src/main/java/com/gildedrose/items/DefaultItem.java
@@ -1,4 +1,6 @@
-package com.gildedrose;
+package com.gildedrose.items;
+
+import com.gildedrose.goblins_grotto.Item;
 
 public class DefaultItem implements Ageable {
